fix typing loc
